processmedia2/pip-requirements.txt
pyyaml
dateutils

pytest
pillow
pytesseract
